implemented make_http_predict_request
