+++ client/python/lib/ovmsclient/tfs_compat/http/requests.py
@@ -14,12 +14,20 @@
 # limitations under the License.
 #
 
+import base64
+import json
+import numpy as np
 from ovmsclient.tfs_compat.base.requests import (PredictRequest, ModelMetadataRequest,
-                                                 ModelStatusRequest)
+                                                 ModelStatusRequest, _check_model_spec)
+from ovmsclient.tfs_compat.grpc.tensors import (NP_TO_TENSOR_MAP, DataType,
+                                                _get_dense_dimensions, _is_bytes_shape_valid)
+from ovmsclient.util.ovmsclient_export import ovmsclient_export
 
 
 class HttpPredictRequest(PredictRequest):
-    pass
+    def __init__(self, inputs, model_name, model_version, parsed_inputs):
+        super().__init__(inputs, model_name, model_version)
+        self.parsed_inputs = parsed_inputs
 
 
 class HttpModelMetadataRequest(ModelMetadataRequest):
@@ -30,6 +38,7 @@ class HttpModelStatusRequest(ModelStatusRequest):
     pass
 
 
+@ovmsclient_export("make_http_predict_request", httpclient="make_predict_request")
 def make_predict_request(inputs, model_name, model_version=0):
     '''
     Create HttpPredictRequest object.
@@ -81,7 +90,24 @@ def make_predict_request(inputs, model_name, model_version=0):
         >>> print(predict_request)
     '''  # noqa: E501
 
-    raise NotImplementedError
+    _check_model_spec(model_name, model_version)
+
+    if not isinstance(inputs, dict):
+        raise TypeError(f'inputs type should be dict, but is {type(inputs).__name__}')
+
+    parsed_inputs = {}
+    for input_name, input_data in inputs.items():
+        if not isinstance(input_name, str):
+            raise TypeError(f'inputs keys should be type str, but found '
+                            f'{type(input_name).__name__}')
+        try:
+            parsed_inputs[input_name] = _parse_input_data(input_data)
+        except Exception as e:
+            print("Error occurred when parsing input: {}. Exited with: {}".format(input_name, e))
+            raise e
+
+    parsed_inputs = json.dumps({"inputs": parsed_inputs})
+    return HttpPredictRequest(inputs, model_name, model_version, parsed_inputs)
 
 
 def make_metadata_request(model_name, model_version=0):
@@ -139,3 +165,39 @@ def make_status_request(model_name, model_version=0):
     '''  # noqa: E501
 
     raise NotImplementedError
+
+
+def _parse_input_data(values):
+
+    # create numpy ndarray from values and find its dtype if not provided
+    if isinstance(values, np.ndarray):
+        tensor_values = values
+    elif isinstance(values, list):
+        dense_dimensions = _get_dense_dimensions(values)
+        tensor_values = np.array(values)
+        if(list(tensor_values.shape) != dense_dimensions):
+            raise ValueError(f'argument must be a dense tensor: {values} - got shape '
+                             f'{list(tensor_values.shape)}, but wanted {dense_dimensions}')
+    elif np.isscalar(values):
+        tensor_values = np.array([values])
+    else:
+        raise TypeError("values type should be (list, np.ndarray, scalar), but is "
+                        f"{type(values).__name__}")
+
+    tensor_type = NP_TO_TENSOR_MAP.get(tensor_values.dtype.type)
+    if tensor_type is not None:
+        dtype = tensor_type.TensorDtype
+    else:
+        raise TypeError("provided values type is not valid")
+
+    if dtype == DataType.DT_STRING and _is_bytes_shape_valid(tensor_values.shape, tensor_values):
+        raise ValueError("bytes values with dtype DT_STRING must be in shape [N]")
+
+    if dtype == DataType.DT_STRING:
+        b64_values = []
+        for value in tensor_values:
+            b64_value = base64.b64encode(value).decode('utf-8')
+            b64_values.append(b64_value)
+        return {"b64": b64_values}
+    else:
+        return tensor_values.tolist()
