+++ client/python/lib/tests/tfs_compat_grpc/test_grpc_responses.py
@@ -18,24 +18,21 @@ from numpy import ndarray
 from numpy.core.numeric import array_equal
 import pytest
 
-from tensorflow.core.protobuf.meta_graph_pb2 import SignatureDef, TensorInfo
-from tensorflow.core.framework.tensor_shape_pb2 import TensorShapeProto
 from tensorflow.core.framework.types_pb2 import DataType
 
-from tensorflow_serving.apis.get_model_metadata_pb2 import GetModelMetadataResponse, SignatureDefMap
-from tensorflow_serving.apis.model_pb2 import ModelSpec
-from tensorflow_serving.apis.get_model_status_pb2 import GetModelStatusResponse, ModelVersionStatus
+from tensorflow_serving.apis.get_model_status_pb2 import ModelVersionStatus
 from tensorflow_serving.apis.predict_pb2 import PredictResponse
-from tensorflow_serving.util.status_pb2 import StatusProto
-
-from google.protobuf.any_pb2 import Any
 
 from ovmsclient.tfs_compat.grpc.responses import (GrpcModelMetadataResponse,
                                                   GrpcModelStatusResponse,
                                                   GrpcPredictResponse)
 
-from config import (MODEL_METADATA_RESPONSE_VALID, MODEL_STATUS_RESPONSE_VALID,
-                    PREDICT_RESPONSE_VALID, PREDICT_RESPONSE_INVALID)
+from tfs_compat_grpc.config import (MODEL_METADATA_RESPONSE_VALID, MODEL_STATUS_RESPONSE_VALID,
+                                    PREDICT_RESPONSE_VALID, PREDICT_RESPONSE_INVALID)
+
+from tfs_compat_grpc.utils import (create_model_metadata_response,
+                                   create_model_status_response,
+                                   merge_model_status_responses)
 
 
 @pytest.mark.parametrize("outputs_dict, model_name, model_version,"
@@ -137,74 +134,3 @@ def test_ModelMetadataResponse_to_dict_valid(model_raw_metadata_response_dict):
                 == model_raw_metadata_response_dict['outputs'][output]['shape'])
         assert (outputs[output]['dtype']
                 == DataType.Name(model_raw_metadata_response_dict['outputs'][output]['dtype']))
-
-
-def create_model_metadata_response(model_raw_metadata_response_dict):
-    raw_response = GetModelMetadataResponse()
-
-    model_spec = ModelSpec()
-    model_spec.name = model_raw_metadata_response_dict['name']
-    model_spec.version.value = model_raw_metadata_response_dict['version']
-
-    raw_response.model_spec.CopyFrom(model_spec)
-
-    signature_def = SignatureDef()
-
-    inputs = model_raw_metadata_response_dict['inputs']
-    for input_name, input_info in inputs.items():
-        input_tensor_info = TensorInfo()
-        input_tensor_info.name = input_name
-        input_tensor_info.dtype = input_info['dtype']
-
-        input_tensor_shape = TensorShapeProto(dim=[
-            TensorShapeProto.Dim(size=x) for x in input_info['shape']
-        ])
-        input_tensor_info.tensor_shape.CopyFrom(input_tensor_shape)
-
-        signature_def.inputs[input_name].CopyFrom(input_tensor_info)
-
-    outputs = model_raw_metadata_response_dict['outputs']
-    for output_name, output_info in outputs.items():
-        output_tensor_info = TensorInfo()
-        output_tensor_info.name = output_name
-        output_tensor_info.dtype = output_info['dtype']
-
-        output_tensor_shape = TensorShapeProto(dim=[
-            TensorShapeProto.Dim(size=x) for x in output_info['shape']
-        ])
-        output_tensor_info.tensor_shape.CopyFrom(output_tensor_shape)
-
-        signature_def.outputs[output_name].CopyFrom(output_tensor_info)
-
-    signature_def_map = SignatureDefMap()
-    signature_def_map.signature_def['serving_default'].CopyFrom(signature_def)
-    serialized_metadata = signature_def_map.SerializeToString()
-
-    metadata = Any()
-    metadata.type_url = "type.googleapis.com/tensorflow.serving.SignatureDefMap"
-    metadata.value = serialized_metadata
-
-    raw_response.metadata['signature_def'].CopyFrom(metadata)
-
-    return raw_response
-
-
-def create_model_status_response(model_version, error_code, error_message, model_state):
-    status = StatusProto()
-    status.error_code = error_code
-    status.error_message = error_message
-
-    model_version_status = ModelVersionStatus()
-    model_version_status.version = model_version
-    model_version_status.state = model_state
-    model_version_status.status.CopyFrom(status)
-
-    return model_version_status
-
-
-def merge_model_status_responses(responses):
-    raw_response = GetModelStatusResponse()
-    model_versions = [response for response in responses]
-    raw_response.model_version_status.extend(model_versions)
-
-    return raw_response
