+++ client/python/lib/tests/tfs_compat_grpc/test_tensors.py
